requirements.txt
aiohttp==3.7.4
aioredis==1.3.1
attrs==20.3.0
biopython==1.78
https://data.pyg.org/whl/torch-1.8.0+cu111/torch_sparse-0.6.12-cp38-cp38-linux_x86_64.whl
https://pytorch-geometric.com/whl/torch-1.8.0+cu111/torch_cluster-1.5.9-cp38-cp38-linux_x86_64.whl
https://pytorch-geometric.com/whl/torch-1.8.0+cu111/torch_scatter-2.0.8-cp38-cp38-linux_x86_64.whl
https://pytorch-geometric.com/whl/torch-1.8.0+cu111/torch_spline_conv-1.2.1-cp38-cp38-linux_x86_64.whl
hydra-core==1.1
numpy==1.19.5
omegaconf==2.1.1
pandas==1.3.4
pip==21.0.1
pre-commit==2.19.0
pylint==2.7.2
pytest==6.2.2
pytest-cov==2.11.1
pytest-order==1.0.0
pytest-sugar==0.9.4
pytest-xdist==2.5.0
pyupgrade==2.29.1
setuptools==60.10.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cu111
torchvision==0.9.0+cu111
torchaudio==0.8.0

--find-links https://pytorch-geometric.com/whl/torch-1.8.0%2Bcu111.html
torch-geometric==2.0.1
